first working test of 20 position components

+++ src_common/component_position.c
@@ -1,7 +1,7 @@
 typedef struct _tComponentTypePosition
 {
-    int x;
-    int y;
+    short int x;
+    short int y;
 } tComponentTypePosition;
 
 void ComponentPositionUpdate(tComponentTypePosition *component)
